spi начало разработки

+++ lib/spi/spi_entity.h
@@ -13,7 +13,7 @@ class SPI
     spi_inst_t *get();
 
   private:
-    static const size_t BAUDRATE = 1000000u;
+    static const uint32_t BAUDRATE = 1000000u;
     spi_inst_t *spi_;
     uint8_t cs_;
